test(Lotto): 단위 테스트 추가
